kv-ee-telegram-bot/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7.0
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.0
httpx==0.25.2
lxml==4.9.3
python-telegram-bot[socks]==20.7.0